+++ e2e/cypress/integration/enterprise/system_console/search_box_spec.js
@@ -8,12 +8,14 @@
 // ***************************************************************
 
 // Stage: @prod
-// Group: @enterprise @system_console
+// Group: @enterprise @not_cloud @system_console
 
 import * as TIMEOUTS from '../../../fixtures/timeouts';
 
 describe('System console', () => {
     before(() => {
+        cy.shouldNotRunOnCloudEdition();
+
         // * Check if server has license
         cy.apiRequireLicense();
     });
